# Micro optimization for matrix multiplication