@MustafaKarabulut 2018-03-04
It looks like Travis has failed because of coding standard checks. I'll fix it as soon as possible.