@akondas 2018-03-04
It looks reasonable, but I prefer use phpbench for performance comparison. There is one good reason: **Retry Threshold**. 

> PHPBench is able to dramatically improve the stability of your benchmarks by retrying the iteration set until all the deviations in time between iterations fit within a given margin of error.
http://phpbench.readthedocs.io/en/latest/benchmark-runner.html#retry-threshold

In this way we can be sure that provided implementation is truly faster.