@marmichalski 2018-03-05
```
+------------------+-------------+-------------+--------+-------+
| subject          | mode        | mean        | rstdev | diff  |
+------------------+-------------+-------------+--------+-------+
| benchMultiply    | 841.972μs   | 888.728μs   | 5.88%  | 1.00x |
| benchOldMultiply | 1,635.211μs | 1,625.570μs | 5.24%  | 1.83x |
+------------------+-------------+-------------+--------+-------+
```

Matrix size: 100x100
Revs: 30
Iterations: 150

I can run the tests with different data if you want.