@MustafaKarabulut 2018-03-05
Thank you @marmichalski. This saved my time to test phpbench which I have never used before.

@akondas It looks like phpbench too confirms the early results.