@akondas 2018-03-05
Yup, you can run `composer fix-cs` and maybe we should leave a comment why such strange code.