@akondas 2018-03-05
With threads we could implement it `core` number times better :rocket: 
Maybe some day ... :calendar: