@MustafaKarabulut 2018-03-08
By the way, I had some issues with the following command for code fixing

`vendor/bin/ecs check src tests --fix`

I had to edit `easy-coding-standard.neon` to remove some dublicates that ecs complained about. Should not it run out-of-the-box as documented? Or am I missing something (again)?